% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{capwords}
\alias{capwords}
\title{Capitalise the first letter of words in a string}
\usage{
capwords(x, strict = FALSE)
}
\arguments{
\item{x}{string}

\item{strict}{logical; FALSE by default. If TRUE, makes all the letters lowercase first before capitalising.}
}
\description{
Capitalise the first letter of words in a string
}
