% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{change.range}
\alias{change.range}
\title{change.range}
\usage{
change.range(object, new_data, ...)
}
\arguments{
\item{object}{obj}

\item{new_data}{new}

\item{...}{dot}
}
\description{
change.range
}
