% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{change.range.simple}
\alias{change.range.simple}
\title{change.range.simple}
\usage{
change.range.simple(data, col_names, min = 0, max = 100)
}
\arguments{
\item{data}{data}

\item{col_names}{names}

\item{min}{num}

\item{max}{num}
}
\description{
change.range.simple
}
