% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{change_range_simple}
\alias{change_range_simple}
\title{Change Range
Adjust the range of variables to the desired minimum and maximum using a simple formula.}
\usage{
change_range_simple(data, col_names, min = 0, max = 100)
}
\arguments{
\item{data}{dataset}

\item{col_names}{names of the columns that need to be adjusted}

\item{min}{numeric; the minimum range; 0 by default.}

\item{max}{numeric; the maximum range; 100 by default.}
}
\value{
the original dataset with the desired columns adjusted
}
\description{
Change Range
Adjust the range of variables to the desired minimum and maximum using a simple formula.
}
