% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjr-journals.R
\name{clean_journal_tables}
\alias{clean_journal_tables}
\title{clean journal tables}
\usage{
clean_journal_tables(df)
}
\arguments{
\item{df}{df}
}
\value{
cleaned df
}
\description{
clean journal tables
}
\examples{
see "OneDrive/PhD Psychology/01 - Investigation - Developing Creditworthiness Measure/01 - Study 1a - Systematic Review/journals.R"

}
