% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{comma}
\alias{comma}
\title{Format a long number with the comma(s)}
\usage{
comma(x, digits = 2)
}
\arguments{
\item{x}{the number (or vector of numbers) to be formatted}

\item{digits}{how many significant digits are to be used; 2 by default. see \code{\link[base:format]{base::format()}}.}
}
\value{
a character (or vector of characters)
}
\description{
Format a long number with the comma(s)
}
