% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{copy_table}
\alias{copy_table}
\title{Copy a dataframe or matrix to clipboard}
\usage{
copy_table(x, row.names = TRUE, col.names = TRUE, clip_size = 2^14, ...)
}
\arguments{
\item{x}{the dataframe or matrix}

\item{row.names}{copy the row names? TRUE by default}

\item{col.names}{copy the column names? TRUE by default}

\item{...}{dots argument passed on to \code{\link[utils:write.table]{utils::write.table()}}}

\item{clipboard_size}{the size limit of the clipboard (in Kb); \code{2^14} by default.
This would only need to be increased if the \code{object.size(x)/1024} exceeds this default.
Please be mindful of the total memory on your machine.}
}
\description{
Copy a dataframe or matrix to clipboard
}
\examples{
\dontrun{
copy_table(mtcars)
}
}
