% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr_simple}
\alias{corr_simple}
\title{corr_simple}
\usage{
corr_simple(data, sig = 0.5, ...)
}
\arguments{
\item{data}{a dataframe}

\item{sig}{numeric; the absolute correlation value above which the correlation is considered notable (.5 by default)}

\item{...}{further arguments for \code{\link[corrplot:corrplot]{corrplot::corrplot()}}}
}
\description{
A small basic function useful for quickly seeing notable correlations.
Good for quick exploration/checking.
Prints a table and produces a plot.
}
