% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-analysis.R
\name{create.nlp.tokens}
\alias{create.nlp.tokens}
\title{create.nlp.tokens}
\usage{
create.nlp.tokens(txt, text_name = "texts")
}
\arguments{
\item{txt}{txt}

\item{text_name}{text_name}
}
\description{
create.nlp.tokens
}
