% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-communicate.R
\name{create.summary.table}
\alias{create.summary.table}
\title{create summary \link[janitor:tabyl]{janitor::tabyl} table}
\usage{
create.summary.table(df, new_names = NULL)
}
\arguments{
\item{df}{dataframe}

\item{new_names}{(optional) a character vector of labels to replace the
column names with as necessary (qs)}
}
\value{
a pretty \code{\link[janitor:tabyl]{janitor::tabyl()}} table
}
\description{
Produces counts, totals, and percentages across all the variables in a dataframe;
with an option to present the variables with different names.
}
\details{
Useful for quickly counting and presenting answers to multiple-choice questions
in a report.
}
