% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-scrape.R
\name{duckduckgo}
\alias{duckduckgo}
\title{duckduckgo search}
\usage{
duckduckgo(searchTerms)
}
\arguments{
\item{searchTerms}{a character vector}
}
\value{
character vector
}
\description{
duckduckgo search
}
