% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{fetch_repos}
\alias{fetch_repos}
\title{fetch_repos}
\usage{
fetch_repos(username, .limit = Inf)
}
\arguments{
\item{username}{github username chr}

\item{.limit}{default Inf}
}
\description{
fetch_repos
}
