% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{fetch_repos_contents}
\alias{fetch_repos_contents}
\title{fetch_repos_contents}
\usage{
fetch_repos_contents(repo_path, .limit = Inf)
}
\arguments{
\item{repo_path}{chr}

\item{.limit}{Inf is defaul}
}
\description{
fetch_repos_contents
}
