% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find.dependencies}
\alias{find.dependencies}
\title{find function dependencies}
\usage{
find.dependencies(e, merge = F)
}
\arguments{
\item{e}{expression}

\item{merge}{TF}
}
\description{
find function dependencies
}
