% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{fit_fun}
\alias{fit_fun}
\title{Will be used to fit the models to different bootstrap data sets}
\usage{
fit_fun(split, ...)
}
\arguments{
\item{split}{s}

\item{...}{dot}
}
\description{
Will be used to fit the models to different bootstrap data sets
}
