% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{foo1}
\alias{foo1}
\title{foo1}
\usage{
foo1(x1, y, x2 = NULL, p = 0.67)
}
\arguments{
\item{x1}{training predictors (dataframe or tibble)}

\item{y}{response variable}

\item{x2}{validation predictors, will be the same as training ones if not provided}

\item{p}{the proportion of the sample to use for training (vs holdout)}
}
\description{
foo1
}
