% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-tidy.R
\name{force_bind}
\alias{force_bind}
\title{Force Bind two Dataframes}
\usage{
force_bind(df1, df2)
}
\arguments{
\item{df1}{dataframe 1}

\item{df2}{dataframe 2}
}
\value{
a dataframe
}
\description{
Force bind (row-wise) two dataframes By forcing their columns to be the same
}
