% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{fuzzy.join}
\alias{fuzzy.join}
\title{Join two tables using fuzzy matching
TODO: not the most accurate implementation for some cases tbh
\itemize{
\item auto join based on what type of join is desired
\item remove the JOIN column at the end
}}
\usage{
fuzzy.join(
  df.x,
  df.y,
  by = NULL,
  max.distance = 0.3,
  ignore.case = TRUE,
  useBytes = TRUE
)
}
\arguments{
\item{df.x}{df}

\item{df.y}{df}

\item{by}{var}

\item{max.distance}{num}

\item{ignore.case}{TF}

\item{useBytes}{TF}
}
\description{
Fuzzy loop
}
