% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{get_poly}
\alias{get_poly}
\title{Generate poly terms of a data matrix / data frame}
\usage{
get_poly(
  xdata = NULL,
  deg = 1,
  maxInteractDeg = deg,
  Xy = NULL,
  model_formula = NULL,
  standardize = FALSE,
  intercept = FALSE
)
}
\arguments{
\item{xdata}{the dataframe (only predictor variables). Factors with more than two levels should not be inputted as integers.}

\item{deg}{the max degree of polynomial terms.}

\item{maxInteractDeg}{the max degree of dummy and nondummy predictor variable interaction terms}

\item{Xy}{the dataframe with the response in the final column (provide xdata or Xy but not both). Factors with more than two levels should not be inputted as integers.}

\item{model_formula}{Internal use. Formula used to generate the training model matrix.
Note: anticipates that polynomial terms are generated using internal functions of library(polyreg)
so YMMV if the formula is not generated on the training data by get_poly().
Also, providing model_formula bypasses deg and maxInteractDeg.}

\item{standardize}{standardize all continuous variables? (Default: FALSE.)}

\item{intercept}{Include intercept? Default: FALSE.}
}
\value{
a model matrix, with the model formula as an additional attribute
}
\description{
Generate poly terms of a data matrix / data frame
}
\examples{
if(FALSE) {
X = get_poly(mtcars, 2)
W = get_poly(Xy=mtcars, 2) # treats final column as response
ncol(W) < ncol(X)          # TRUE
X_train <- get_poly(mtcars[1:20,], 4, 2)
X_test <- get_poly(mtcars[21:32,], model_formula = attributes(X_train)$formula)
}
}
