% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{get_values}
\alias{get_values}
\title{Get Values from Data}
\usage{
get_values(case, include_na = TRUE)
}
\arguments{
\item{case}{Which case you want extract data from?}

\item{include_na}{Should NA be an option? `TRUE` by default}
}
\description{
These functions extract sets of possible values for
columns within the data
}
