% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-creation.R
\name{git_setup}
\alias{git_setup}
\title{git_setup (quite experimental atm)}
\usage{
git_setup(
  repo_name = NULL,
  private = TRUE,
  user = gh::gh_whoami()$login,
  git_ignore = NULL
)
}
\arguments{
\item{repo_name}{character vector}

\item{private}{T/F}

\item{user}{default uses gh::gh_whoami()$"login"}

\item{git_ignore}{additional folders or files to pass onto .gitignore.}
}
\description{
git_setup (quite experimental atm)
}
