% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{glmnet_cor}
\alias{glmnet_cor}
\title{Function to create and validate elastic net prediction models}
\usage{
glmnet_cor(
  x1,
  y,
  x2 = NULL,
  prop = 0.67,
  alpha = 0.05,
  lambda.type = "lambda.min"
)
}
\arguments{
\item{x1}{training predictors (dataframe or tibble)}

\item{y}{response variable}

\item{x2}{validation predictors, will be the same as training ones if not provided}

\item{prop}{the proportion of the sample to use for training (vs holdout)}

\item{alpha}{passed to cv.glmnet; default is .05}

\item{lambda.type}{passed to cv.glmnet; default is "lambda.min"}
}
\description{
Function to create and validate elastic net prediction models
}
