% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-tidy.R
\name{is.count}
\alias{is.count}
\title{check if count}
\usage{
is.count(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{a vector}

\item{tol}{tolerance (e.g., 0.000000015); is .Machine$double.eps^0.5 by default}
}
\value{
logical value (TRUE/FALSE)
}
\description{
check if count
}
