% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-tidy.R
\name{isOrd}
\alias{isOrd}
\title{Identify ordinal variables}
\usage{
isOrd(data, n = 7)
}
\arguments{
\item{data}{dataset}

\item{n}{max number of levels allowed for the variable before it stops being
considered ordinal and becomes continuous. Default is 7.}
}
\description{
Identify ordinal variables
}
