% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{letters_to_numbers}
\alias{letters_to_numbers}
\title{Letters to Numbers
Takes a character vector of letters from the Roman alphabet
and transforms them into their numeric equivalent value.
The case (lower vs upper) is irrelevant.
Returns \code{NA}, if the letter is not part of the Roman alphabet.
Useful little tool for basic cipher challenges.}
\usage{
letters_to_numbers(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
interger vector
}
\description{
Letters to Numbers
Takes a character vector of letters from the Roman alphabet
and transforms them into their numeric equivalent value.
The case (lower vs upper) is irrelevant.
Returns \code{NA}, if the letter is not part of the Roman alphabet.
Useful little tool for basic cipher challenges.
}
\examples{
letters_to_numbers("a")
letters_to_numbers(c("p", "n"))
}
