% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-scrape.R
\name{link_to_df}
\alias{link_to_df}
\title{Given a relative wiki link, output a tibble with the pages it links to}
\usage{
link_to_df(relative_link)
}
\description{
Given a relative wiki link, output a tibble with the pages it links to
}
\keyword{internal}
