% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{my.lasso.function}
\alias{my.lasso.function}
\title{My LASSO Function}
\usage{
my.lasso.function(
  data,
  x,
  y,
  nrep = 100,
  family = "gaussian",
  alpha = 1,
  lambda = NULL,
  lambda.min = TRUE
)
}
\arguments{
\item{data}{the dataset}

\item{x}{x (character)}

\item{y}{y (character)}

\item{nrep}{number of repetitions}

\item{family}{type; default is 'gaussian'}

\item{alpha}{alpha value for LASSO; default is 1}

\item{lambda}{provide own matrix of lambdas to test or use the def}

\item{lambda.min}{TRUE/FALSE; if FALSE lambda.1se is used instead}
}
\description{
My LASSO Function
}
