% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natalia-package.R
\docType{package}
\name{natalia-package}
\alias{natalia}
\alias{natalia-package}
\title{natalia: A Package to Support Natalia in her Data Science Shenanigans}
\description{
Bits & Bobs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natalia-mlad/natalia}
  \item Report bugs at \url{https://github.com/natalia-mlad/natalia/issues}
}

}
\author{
\strong{Maintainer}: Natalia Mladentseva \email{natashka.ml@gmail.com}

}
\keyword{internal}
