% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-creation.R
\name{natalia_use_r}
\alias{natalia_use_r}
\title{natalia_use_r}
\usage{
natalia_use_r(name = NULL, open = rlang::is_interactive())
}
\arguments{
\item{name}{NULL}

\item{open}{rlang::is_interactive()}
}
\description{
natalia_use_r
}
