% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{polytomize_data}
\alias{polytomize_data}
\title{polytomize data to demo ordinal models}
\usage{
polytomize_data(x)
}
\arguments{
\item{x}{?a vector?}
}
\description{
polytomize data to demo ordinal models
}
