% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_pandas.R
\name{python_df}
\alias{python_df}
\title{Takes in an unconverted pandas.DataFrame via \code{reticulate::py_eval} and returns
an R data.frame that has the look of a pandas.DataFrame}
\usage{
python_df(pydf)
}
\arguments{
\item{pydf}{a python dataframe}
}
\value{

}
\description{
Takes in an unconverted pandas.DataFrame via \code{reticulate::py_eval} and returns
an R data.frame that has the look of a pandas.DataFrame
}
