% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{quick_git}
\alias{quick_git}
\title{Quick Git
Stage all, commit, and push to github all with one function.}
\usage{
quick_git(commit_msg = NULL)
}
\arguments{
\item{commit_msg}{Git Commit Message}
}
\description{
Quick Git
Stage all, commit, and push to github all with one function.
}
