% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-communicate.R
\name{quick_summary}
\alias{quick_summary}
\title{quick_summary}
\usage{
quick_summary(df, variable, desc = TRUE)
}
\arguments{
\item{df}{the dataframe}

\item{variable}{the var}

\item{desc}{logical. default TRUE}
}
\value{
summarised df
}
\description{
quick_summary
}
