% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-analysis.R
\name{remove_hyphen}
\alias{remove_hyphen}
\title{remove_hyphen - from pdfsearch function}
\usage{
remove_hyphen(text_lines)
}
\arguments{
\item{text_lines}{text_lines}
}
\description{
remove_hyphen - from pdfsearch function
}
