% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-tidy.R
\name{remove_identical_cols}
\alias{remove_identical_cols}
\title{remove_identical_cols
Tested to be the fastest out of other possible methods!
aka fun4}
\usage{
remove_identical_cols(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
a tidier dataframe
}
\description{
remove_identical_cols
Tested to be the fastest out of other possible methods!
aka fun4
}
