% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-vis.R
\name{rocplot}
\alias{rocplot}
\title{Plot a ROC curve}
\usage{
rocplot(pred, truth, ...)
}
\arguments{
\item{pred}{predicted value}

\item{truth}{actual value}

\item{...}{additional arguments passed on to ROCR::plot-methods}
}
\description{
Plot a ROC curve
}
