% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web-scrape.R
\name{scrapeurl}
\alias{scrapeurl}
\title{scrape url}
\usage{
scrapeurl(my_url, path)
}
\arguments{
\item{my_url}{url of the head/start page}

\item{path}{dir path where to save the html files}
}
\value{
a character vector of all the urls scraped
}
\description{
scrape url
}
\examples{
# scrapeurl("https://loal.app/", path_home("OneDrive/PhD Psychology/loal-website-scrape"))
# "https://www.oakam.com/" \%>\% scrapeurl(path_home("OneDrive/PhD Psychology/oakam-website-scrape"))

}
