% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-api.R
\name{search_github}
\alias{search_github}
\title{search_github}
\usage{
search_github(my_query, max = 100, wait = 10)
}
\arguments{
\item{my_query}{chr query; see https://docs.github.com/en/search-github/searching-on-github/searching-code
filename: path: language: size: extension:}

\item{max}{number of results; 100 default}

\item{wait}{how many seconds minimum to wait between requests; default 10}
}
\value{
tibble
}
\description{
search_github
}
