% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{simulate_data_continuous}
\alias{simulate_data_continuous}
\title{Simulate data from the model.}
\usage{
simulate_data_continuous()
}
\value{
A data frame with the following columns.
\itemize{
\item \code{y}: Simulated normal responses.
\item \code{x}: A simulated covariate of zeroes and ones.
\item \code{beta_true}: The value of the regression coefficient \code{beta}
used to simulate the data.
}
}
\description{
Use a continuous covariate x.
}
\examples{
if(FALSE) {
simulate_data_continuous()
}

}
