% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{spearman_brown}
\alias{spearman_brown}
\title{spearman_brown}
\usage{
spearman_brown(df, items, name = "", SB_only = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{items}{items}

\item{name}{character}

\item{SB_only}{TRUE/FALSE; default is false}
}
\description{
spearman_brown
}
