% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{theme_apa}
\alias{theme_apa}
\title{APA Themed Flextable}
\usage{
theme_apa(x, size = 10)
}
\arguments{
\item{x}{a flextable}

\item{size}{font size for the table (default = 10)}
}
\value{
a flextable
}
\description{
A theme function is a R function where input is a flextable and output is a flextable.
You are free to do any formatting inside this function.
}
\note{
https://github.com/davidgohel/flextable/blob/master/R/themes.R
}
\examples{
if(FALSE) {
ft <- flextable::flextable(head(cars))
ft <- theme_apa(ft)
ft
}

}
