% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-tidy.R, R/ds-transform.R
\name{tidy_my_data}
\alias{tidy_my_data}
\title{Tidy my Data}
\usage{
tidy_my_data(data)

tidy_my_data(data)
}
\arguments{
\item{data}{dataframe}

\item{na.rm}{T/F}

\item{quiet}{T/F}
}
\value{
a tidier dataframe

tidier data
}
\description{
Tidy my Data

tidy_my_data
}
