% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-scrape.R
\name{wiki_grab_links}
\alias{wiki_grab_links}
\title{grab_wiki_links}
\usage{
wiki_grab_links(url, keep_duplicates = FALSE)
}
\arguments{
\item{url}{character. the url of the wiki page}

\item{keep_duplicates}{logical.}
}
\value{

}
\description{
grab_wiki_links
}
