% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-scrape.R
\name{wiki_link_to_title}
\alias{wiki_link_to_title}
\title{Convert a wikipedia link to the title}
\usage{
wiki_link_to_title(wiki_link)
}
\description{
Convert a wikipedia link to the title
}
\keyword{internal}
