% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{zero}
\alias{zero}
\title{zero
Rounds to two decimal places (by default) and adds trailing zero(es) which are normally omitted}
\usage{
zero(x, k = 2)
}
\arguments{
\item{x}{x}

\item{k}{number of decimals (2 by default)}
}
\value{
character vector
}
\description{
zero
Rounds to two decimal places (by default) and adds trailing zero(es) which are normally omitted
}
