% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{copy_table}
\alias{copy_table}
\title{Copy a dataframe or matrix to clipboard}
\usage{
copy_table(x, row.names = TRUE, col.names = TRUE, ...)
}
\arguments{
\item{x}{the dataframe or matrix}

\item{row.names}{copy the row names? TRUE by default}

\item{col.names}{copy the column names? TRUE by default}

\item{...}{dots argument}
}
\description{
Copy a dataframe or matrix to clipboard
}
