% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-communicate.R
\name{create.summary.table}
\alias{create.summary.table}
\title{create.summary.table}
\usage{
create.summary.table(df, new_names = NULL)
}
\arguments{
\item{df}{dataframe}

\item{new_names}{(optional) a character vector of labels to replace the
column names with as necessary (qs)}
}
\value{
a pretty tabyl
}
\description{
create.summary.table
}
