% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find.dependencies}
\alias{find.dependencies}
\title{find a function's dependencies}
\usage{
find.dependencies(e, merge = FALSE)
}
\arguments{
\item{e}{an expression}

\item{merge}{logical; FALSE by default; passed onto \link[codetools:findGlobals]{codetools::findGlobals}}
}
\description{
find a function's dependencies
}
