% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{flat_cor_mat}
\alias{flat_cor_mat}
\title{Simple formatting of a correlation matrix}
\usage{
flat_cor_mat(cor_r, cor_p = NA)
}
\arguments{
\item{cor_r}{correlation matrix with r values}

\item{cor_p}{(optional) matrix of correlation p values}
}
\value{
Column 1 : row names (variable 1 for the correlation test)
Column 2 : column names (variable 2 for the correlation test)
Column 3 : the correlation coefficients
Column 4 : the p-values of the correlations
}
\description{
This function provides a simple formatting of a correlation matrix into a table with 4 columns containing:
Column 1 : row names (variable 1 for the correlation test)
Column 2 : column names (variable 2 for the correlation test)
Column 3 : the correlation coefficients
Column 4 : the p-values of the correlations
}
