% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{fuzzy.join}
\alias{fuzzy.join}
\title{Join two tables using fuzzy matching}
\usage{
fuzzy.join(
  df.x,
  df.y,
  by = NULL,
  max.distance = 0.3,
  ignore.case = TRUE,
  useBytes = TRUE
)
}
\description{
Fuzzy loop
}
