% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{letters_to_numbers}
\alias{letters_to_numbers}
\title{letters_to_numbers}
\usage{
letters_to_numbers(x)
}
\arguments{
\item{x}{character vector}
}
\value{
interger vector
}
\description{
letters_to_numbers
}
\examples{
letters_to_numbers("a")
letters_to_numbers(c("p", "n"))
}
