% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{LoadToEnvironment}
\alias{LoadToEnvironment}
\title{Load data to a separate environment}
\usage{
LoadToEnvironment(RData, env = new.env())
}
\arguments{
\item{RData}{path for the .RData file}

\item{env}{new.env}
}
\description{
Load data to a separate environment
}
\examples{
if(FALSE) {
march2021.data <- LoadToEnvironment("March 2021.RData")
dataset <- march2021.data$dataset
}
}
