% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{nozero}
\alias{nozero}
\title{removes leading zero before decimals and rounds number to two decimals}
\usage{
nozero(x, k = 2)
}
\arguments{
\item{x}{x}

\item{k}{number of decimals}
}
\description{
removes leading zero before decimals and rounds number to two decimals
}
