% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{nozero}
\alias{nozero}
\title{nozero
Removes leading zero before decimals and rounds number to two decimals (by default)}
\usage{
nozero(x, k = 2)
}
\arguments{
\item{x}{numeric vector}

\item{k}{number of decimals (2 by default)}
}
\value{
character vector
}
\description{
nozero
Removes leading zero before decimals and rounds number to two decimals (by default)
}
