% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{polytomize_data}
\alias{polytomize_data}
\title{Polytomize Data
polytomize data to, for example, demo ordinal models}
\usage{
polytomize_data(df, thresholds)
}
\arguments{
\item{df}{a dataframe where all numeric}

\item{thresholds}{a numeric vector of cutoff points for the
category thresholds; provide one fewer than the required number
of categories (e.g., c(-0.5, 1) for three categories)}
}
\description{
Polytomize Data
polytomize data to, for example, demo ordinal models
}
