% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-creation.R
\name{quick_git}
\alias{quick_git}
\title{quick git stage -> commit -> push}
\usage{
quick_git(commit_msg = NULL)
}
\arguments{
\item{commit_msg}{Git Commit Message}
}
\description{
quick git stage -> commit -> push
}
