% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-tidy.R
\name{remove_identical_cols}
\alias{remove_identical_cols}
\title{Remove Identical Cols}
\usage{
remove_identical_cols(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
a tidier dataframe
}
\description{
Remove Identical Cols
}
\details{
Tested to be the fastest out of other possible methods! aka fun4
}
