% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-transform.R
\name{zero}
\alias{zero}
\title{formats numbers to no more or less than two decimals}
\usage{
zero(x, k = 2)
}
\arguments{
\item{x}{x}

\item{k}{number of decimals}
}
\description{
formats numbers to no more or less than two decimals
}
